+++ changelog.md
@@ -5,10 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.7] - 2024-08-20
+### Added
+-Added __str__ to board.py 
 
 ## [0.0.6] - 2024-08-19
 ### Added
--Added reina, rey and peones to board.py and validate coordinate to
+-Added reina, rey and peones to board.py and validate coordinate to board.py 
+
 ## [0.0.5] - 2024-08-18
 ### Added
 -Added test_piece.py with some mistakes y one of the test
@@ -24,7 +28,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.2] - 2024-08-15
 ### Added
 - Incorporation of hourse token to the board.
-
 
 ## [0.0.1] - 2024-08-14
 ### Added